Move all the scene stuff into the view
